Separate classes into their respective files
