+++ nodeServer/views/dataManagement.pug
@@ -5,7 +5,7 @@ html
     script.
       window.dataManagementConfigFromServer = !{dataManagementConfig};
 
-    script(type="module" src="/javascripts/dataManagement.js")
+    script(type="module" src="/javascripts/data-management/dataManagementInitialization.js")
 
     link(rel='stylesheet', href='/stylesheets/style.css')
